[Playground] Resolve CI/CD failures
Fetch example execution results right after each example execution instead of waiting for all examples to finish execution
